pyproject.toml
[project]
name = "memo-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "chromadb>=1.4.0",
    "colorama>=0.4.6",
    "datasets>=4.4.2",
    "faiss-cpu>=1.13.2",
    "ffmpeg>=1.4",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-genai>=1.57.0",
    "google-generativeai>=0.8.6",
    "gunicorn>=24.0.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.1.3",
    "librosa>=0.11.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "soundfile>=0.13.1",
    "streamlit>=1.53.1",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
